fix overwrite of savedata when no import file is present on usb device

+++ toxblinkenwall/process_usb_drive.sh
@@ -157,8 +157,8 @@ function export_data_toxsave
 
     mkdir -p "$mount_dir""/""backup/"
     echo "backing up:Tox Privkey and Friends" >> "$logfile"
-    mv -v "$mount_dir""/""backup/"savedata.tox.1 "$mount_dir""/""backup/"savedata.tox.2 >> "$logfile" 2>&1
-    mv -v "$mount_dir""/""backup/"savedata.tox "$mount_dir""/""backup/"savedata.tox.1 >> "$logfile" 2>&1
+    cp -v "$mount_dir""/""backup/"savedata.tox.1 "$mount_dir""/""backup/"savedata.tox.2 >> "$logfile" 2>&1
+    cp -v "$mount_dir""/""backup/"savedata.tox "$mount_dir""/""backup/"savedata.tox.1 >> "$logfile" 2>&1
     cp -v "$dst_dir""/"savedata.tox "$mount_dir""/""backup/"savedata.tox >> "$logfile" 2>&1
 
     return 0
@@ -230,19 +230,23 @@ function pair_usb_device
 
 function import_data_toxsave
 {
-    # tell ToxBlinkenwall to restart
-    if [ -p "/home/pi/ToxBlinkenwall/toxblinkenwall/ext_keys.fifo" ]; then
-        echo "restart:" >> /home/pi/ToxBlinkenwall/toxblinkenwall/ext_keys.fifo
+    if [ -f "$mount_dir""/"savedata.tox ]; then
+        # tell ToxBlinkenwall to restart
+        if [ -p "/home/pi/ToxBlinkenwall/toxblinkenwall/ext_keys.fifo" ]; then
+            echo "restart:" >> /home/pi/ToxBlinkenwall/toxblinkenwall/ext_keys.fifo
+        fi
+
+        sleep 3 # wait for ToxBlinkenwall to shutdown (it will write toxsave on shudown!) # TODO: make more exact!
+
+        # overwrite toxsave with the backup
+        cp -v "$dst_dir""/"savedata.tox.BCK.1 "$dst_dir""/"savedata.tox.BCK.2 >> "$logfile" 2>&1 # just in case, we make a copy. you never know :-)
+        cp -v "$dst_dir""/"savedata.tox "$dst_dir""/"savedata.tox.BCK.1 >> "$logfile" 2>&1 # just in case, we make a copy. you never know :-)
+        cp -v "$mount_dir""/"savedata.tox "$dst_dir""/"savedata.tox >> "$logfile" 2>&1
+
+        return 0
     fi
 
-    sleep 3 # wait for ToxBlinkenwall to shutdown (it will write toxsave on shudown!) # TODO: make more exact!
-
-    # overwrite toxsave with the backup
-    mv -v "$dst_dir""/"savedata.tox.BCK.1 "$dst_dir""/"savedata.tox.BCK.2 >> "$logfile" 2>&1 # just in case, we make a copy. you never know :-)
-    mv -v "$dst_dir""/"savedata.tox "$dst_dir""/"savedata.tox.BCK.1 >> "$logfile" 2>&1 # just in case, we make a copy. you never know :-)
-    cp -v "$mount_dir""/"savedata.tox "$dst_dir""/"savedata.tox >> "$logfile" 2>&1
-
-    return 0
+    return 1
 }
 
 #
